refactor: 리뷰 관련 폴더 구조 개선
[#435]
