+++ frontend/src/components/ui/StationDetailsWindow/reviews/ReviewList.tsx
@@ -10,7 +10,7 @@ import Text from '@common/Text';
 
 import ReviewCard from '@ui/StationDetailsWindow/reviews/ReviewCard';
 import ReviewCardsLoading from '@ui/StationDetailsWindow/reviews/ReviewCardsLoading';
-import ReviewCreate from '@ui/StationDetailsWindow/reviews/ReviewCreate';
+import ReviewCreate from '@ui/StationDetailsWindow/reviews/crud/ReviewCreate';
 
 import type { Review } from '@type';
