+++ frontend/src/components/ui/StationDetailsWindow/reviews/previews/UserRatings.tsx
